move `TLSConfig` into separate file
